Basic Tagging System Working & simplistic private bookmarks system
Managed to get a simple tagging system functioning. It currently allows for users to create bookmarks with tags and view the bookmark details alongside a list of associated tags. The user is then able to select a tag and view all bookmarks associated with that tag that they have permission to view.